requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Authlib
requests
python-dotenv
Pillow
qrcode
reportlab
gunicorn
psycopg2-binary